@@ -1,5 +1,5 @@
-name: money_trail
-description: A new Flutter project.
+name: fowallet
+description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.3 <4.0.0'
+  sdk: '>=3.2.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,17 +32,20 @@ dependencies:
     sdk: flutter
 
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  get: ^4.6.5
   infinite_carousel: ^1.0.3
-  fl_chart: ^0.65.0
-#  firebase_auth: ^4.7.1
-#  firebase_core: ^2.15.0
-#  cloud_firestore: ^4.8.3
+  firebase_auth: ^4.15.3
+  firebase_core: ^2.15.0
+  get: ^4.6.5
+  cloud_firestore: ^4.13.6
   font_awesome_flutter: ^10.5.0
+  fl_chart: ^0.65.0
+  intl: ^0.19.0
+  google_sign_in: ^6.2.1
+  googleapis: ^12.0.0
+  googleapis_auth: ^1.4.2
 
 dev_dependencies:
   flutter_test:
@@ -82,15 +85,14 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
-  - family: Titillium Web Regular
-    fonts:
-      - asset: fonts/TitilliumWeb-Regular.ttf
-  - family: Titillium Web Bold
-    fonts:
-      - asset: fonts/TitilliumWeb-Bold.ttf
-  - family: Titillium Web SemiBold
-    fonts:
-      - asset: fonts/TitilliumWeb-SemiBold.ttf
-
+    - family: Titillium Web Regular
+      fonts:
+        - asset: fonts/TitilliumWeb-Regular.ttf
+    - family: Titillium Web Bold
+      fonts:
+        - asset: fonts/TitilliumWeb-Bold.ttf
+    - family: Titillium Web SemiBold
+      fonts:
+        - asset: fonts/TitilliumWeb-SemiBold.ttf
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
