@@ -1,14 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:get/get.dart';
+import 'package:money_trail/model/model_transaction.dart';
+import 'package:money_trail/screens/expense_screen/AddExpenseController.dart';
 import 'package:money_trail/utils/constant/dime.dart';
 import 'package:money_trail/utils/styles/boder_style.dart';
 import 'package:money_trail/utils/styles/text_styles.dart';
 
 class FormItem extends StatelessWidget {
-  const FormItem({Key? key, this.title, required this.icon, this.hint}) : super(key: key);
+  FormItem({super.key, this.title, required this.icon, this.hint, required this.modelTransactionObj});
   final hint;
   final title;
   final icon;
+  ModelTransaction modelTransactionObj;
+
+  final TextEditingController textEditingController = TextEditingController();
+  // AddExpenseController addExpenseController = Get.put(AddExpenseController());
 
   @override
   Widget build(BuildContext context) {
@@ -22,6 +29,11 @@ class FormItem extends StatelessWidget {
           Padding(
             padding: const EdgeInsets.symmetric(vertical: DimeUtil.sm),
             child: TextField(
+              onChanged: (changeContent){
+                setTransactionData(title, changeContent);
+                print('Rs ${changeContent}');
+              },
+              controller: textEditingController,
                 decoration: InputDecoration(
                     focusedBorder: BorderUtilStyle.getFocusBorderStyle(),
                     border: BorderUtilStyle.getBorderStyle(),
@@ -36,4 +48,37 @@ class FormItem extends StatelessWidget {
       ),
     );
   }
+
+  setTransactionData(String key, var value){
+    switch(key) {
+      case "TITLE" :
+        modelTransactionObj.title = value;
+        break;
+      case "DATE" :
+        modelTransactionObj.date = value;
+        break;
+      case "TYPE" :
+        modelTransactionObj.type = value;
+        break;
+      case "EXTRA01" :
+        modelTransactionObj.extra01 = value;
+        break;
+      case "EXTRA02" :
+        modelTransactionObj.extra02 = value;
+        break;
+      case "EXTRA03" :
+        modelTransactionObj.extra03 = value;
+        break;
+      case "EXTRA04" :
+        modelTransactionObj.extra04 = value;
+        break;
+      case "EXTRA05" :
+        modelTransactionObj.extra05 = value;
+        break;
+      case "SUM" :
+        modelTransactionObj.amount = double.parse(value);
+        break;
+    }
+  }
 }
+
