@@ -1,3 +1,4 @@
+import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:money_trail/utils/constant/dime.dart';
@@ -7,43 +8,25 @@ class BudgetPlaner extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListView.builder(
-        itemCount: 6,
-        itemBuilder: (context, _) => Container(
-              padding: const EdgeInsets.all(DimeUtil.sm),
-              margin: const EdgeInsets.all(DimeUtil.sm),
-              child: Column(
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Row(
-                        children: [
-                          Padding(
-                            padding: EdgeInsets.all(8.0),
-                            child: FaIcon(FontAwesomeIcons.shirt),
-                          ),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Clothes"),
-                              Text("Rs 15,000"),
-
-                            ],
-                          )
-                        ],
-                      ),
-                    ],
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(DimeUtil.sm),
-                    child: LinearProgressIndicator(
-                      color: Colors.grey,
-                      value: 40.0,
-                    ),
-                  )
-                ],
-              ),
-            ));
+    return PieChart(
+        PieChartData(sectionsSpace: 0.0, centerSpaceRadius: 0.0, sections: [
+      PieChartSectionData(
+          value: 3.0,
+          color: Colors.green,
+          radius: 100,
+          badgeWidget: Container(
+              decoration: BoxDecoration(
+                  borderRadius:
+                      BorderRadius.circular(DimeUtil.borderFullCircle),
+                  color: Colors.black),
+              padding: const EdgeInsets.all(DimeUtil.md),
+              child: const FaIcon(
+                FontAwesomeIcons.bowlFood,
+                color: Colors.white,
+              )),
+          badgePositionPercentageOffset: 1.0,),
+      PieChartSectionData(value: 1.0, color: Colors.red, radius: 100),
+      PieChartSectionData(value: 4.0, color: Colors.orangeAccent, radius: 100),
+    ]));
   }
 }
