@@ -1,1 +1,14 @@
-const primaryColorHex = "#11D05E";
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
+
+class ColorsUtil{
+  // HEX : "#01784E"
+  // static const primaryColorHex =  Color.fromRGBO(1, 120, 78, 1);
+  static const primaryColor =  Colors.teal;
+  static const secondaryColor =  Colors.red;
+  static const baseBlackColor =  Colors.black;
+  static const baseWhiteColor =  Colors.white;
+  static Color? baseGreyColor =  Colors.grey[800];
+
+}
